Extract duplicate meta data providers to trait.
